backend/requirements/base.txt
# ── Framework ────────────────────────────
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0

# ── Auth ─────────────────────────────────
djangorestframework-simplejwt>=5.3,<6.0

# ── Database ─────────────────────────────
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0

# ── Geospatial ───────────────────────────
# PostGIS support is built into Django (django.contrib.gis)

# ── Async Tasks ──────────────────────────
celery[redis]>=5.4,<6.0
django-celery-beat>=2.6,<3.0

# ── API Documentation ────────────────────
drf-spectacular>=0.27,<1.0

# ── Scraping ─────────────────────────────
scrapy>=2.11,<3.0
playwright>=1.44,<2.0
scrapy-playwright>=0.0.34,<1.0

# ── AI / ML ──────────────────────────────
google-genai>=0.8,<2.0
ortools>=9.10,<10.0
thefuzz[speedup]>=0.22,<1.0
pypdf>=4.2,<6.0

# ── Utilities ────────────────────────────
python-decouple>=3.8,<4.0
structlog>=24.0,<25.0
sentry-sdk[django]>=2.8,<3.0
Pillow>=10.4,<11.0
requests>=2.32,<3.0
redis>=5.0,<6.0
exponent_server_sdk>=2.0,<3.0
